MessageTextViewをユーザ用とキャラクタ用の２つのクラスに分けた
