--- HackAichi2024/HackAichi2024/MessageTextView.swift
@@ -1,41 +1,0 @@
-//
-//  MessageTextView.swift
-//  HackAichi2024
-//
-//  Created by jun on 2024/09/15.
-//
-
-import UIKit
-
-enum MessageSender {
-    case myself
-    case character
-}
-
-class MessageTextView: UITextView {
-    private var messageSender: MessageSender
-    private var textViewBackgroundColor: UIColor {
-        switch messageSender {
-        case .myself:
-            return .systemGreen
-        case .character:
-            return .systemGray5
-        }
-    }
-
-    init(sender: MessageSender) {
-        self.messageSender = sender
-        super.init(frame: .zero, textContainer: nil)
-        setUpTextView()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setUpTextView() {
-        self.backgroundColor = textViewBackgroundColor
-        self.layer.cornerRadius = 20
-        self.translatesAutoresizingMaskIntoConstraints = false
-    }
-}
